bgls: Add documentation, minor refactor, fix linting, fix Kosk
This commit adds the documentation to the various bgls
methods. Method names have been refactored, to make it more clear
when they should be used. Additionally it makes the entire codebase
pass gometalinter.

This commit also removes the possibility for a rogue public key attack
with Kosk. See blsKosk.go for the new scheme.